moves main file
